serenity-reporter/requirements.txt
coinbasepro==0.2.0
fire==0.2.1
gemini-python==0.2.0
